# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Capability": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "blocked",
              "unavailable",
              "unverified"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status",
          "detail"
        ],
        "title": "Capability",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "capability_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Capability Digest",
            "type": "string"
          },
          "readiness": {
            "enum": [
              "ready",
              "degraded",
              "blocked"
            ],
            "title": "Readiness",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "readiness",
          "capability_digest",
          "capabilities"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "KnowledgeKind": {
        "enum": [
          "observed",
          "reported",
          "inferred",
          "verified"
        ],
        "title": "KnowledgeKind",
        "type": "string"
      },
      "PairRequest": {
        "additionalProperties": false,
        "properties": {
          "nonce": {
            "maxLength": 256,
            "minLength": 32,
            "title": "Nonce",
            "type": "string"
          }
        },
        "required": [
          "nonce"
        ],
        "title": "PairRequest",
        "type": "object"
      },
      "RunEvidenceEventView": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "id": {
            "pattern": "^[A-Za-z0-9_-]{1,80}$",
            "title": "Id",
            "type": "string"
          },
          "knowledge_kind": {
            "$ref": "#/components/schemas/KnowledgeKind"
          },
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seq",
          "title",
          "detail",
          "knowledge_kind"
        ],
        "title": "RunEvidenceEventView",
        "type": "object"
      },
      "RunEvidenceView": {
        "additionalProperties": false,
        "properties": {
          "changed_paths": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Changed Paths",
            "type": "array"
          },
          "codex_thread_id": {
            "pattern": "^[A-Za-z0-9_-]{1,80}$",
            "title": "Codex Thread Id",
            "type": "string"
          },
          "codex_turn_id": {
            "pattern": "^[A-Za-z0-9_-]{1,80}$",
            "title": "Codex Turn Id",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvidenceEventView"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          },
          "final_snapshot_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Final Snapshot Sha256",
            "type": "string"
          },
          "initial_snapshot_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Initial Snapshot Sha256",
            "type": "string"
          },
          "model_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "project": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Project",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "release_commit": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Release Commit",
            "type": "string"
          },
          "run_id": {
            "pattern": "^[A-Za-z0-9_-]{1,80}$",
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "const": "live_receipt",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "const": "verified",
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "verification": {
            "$ref": "#/components/schemas/RunVerificationView"
          }
        },
        "required": [
          "run_id",
          "title",
          "project",
          "status",
          "source",
          "recorded_at",
          "release_commit",
          "model_id",
          "codex_thread_id",
          "codex_turn_id",
          "initial_snapshot_sha256",
          "final_snapshot_sha256",
          "changed_paths",
          "verification",
          "events"
        ],
        "title": "RunEvidenceView",
        "type": "object"
      },
      "RunVerificationView": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "output_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Output Sha256",
            "type": "string"
          },
          "snapshot_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Snapshot Sha256",
            "type": "string"
          },
          "status": {
            "const": "passed",
            "title": "Status",
            "type": "string"
          },
          "suite_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Suite Digest",
            "type": "string"
          }
        },
        "required": [
          "status",
          "snapshot_sha256",
          "suite_digest",
          "output_sha256",
          "duration_ms"
        ],
        "title": "RunVerificationView",
        "type": "object"
      },
      "UsageAdviceView": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_requested": {
            "title": "Checkpoint Requested",
            "type": "boolean"
          },
          "level": {
            "enum": [
              "normal",
              "prepare",
              "pause",
              "critical",
              "unknown"
            ],
            "title": "Level",
            "type": "string"
          },
          "pause_managed_run_requested": {
            "title": "Pause Managed Run Requested",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "level",
          "checkpoint_requested",
          "pause_managed_run_requested",
          "reason"
        ],
        "title": "UsageAdviceView",
        "type": "object"
      },
      "UsageResponse": {
        "additionalProperties": false,
        "properties": {
          "advice": {
            "$ref": "#/components/schemas/UsageAdviceView"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "source": {
            "const": "codex-app-server",
            "title": "Source",
            "type": "string"
          },
          "state": {
            "enum": [
              "known",
              "unknown",
              "stale",
              "unavailable"
            ],
            "title": "State",
            "type": "string"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/UsageWindowView"
            },
            "title": "Windows",
            "type": "array"
          }
        },
        "required": [
          "observed_at",
          "state",
          "windows",
          "source",
          "error",
          "advice"
        ],
        "title": "UsageResponse",
        "type": "object"
      },
      "UsageWindowView": {
        "additionalProperties": false,
        "properties": {
          "limit_id": {
            "title": "Limit Id",
            "type": "string"
          },
          "remaining_percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Remaining Percent",
            "type": "number"
          },
          "resets_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resets At"
          },
          "used_percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Used Percent",
            "type": "number"
          },
          "window": {
            "enum": [
              "primary",
              "secondary"
            ],
            "title": "Window",
            "type": "string"
          },
          "window_minutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Minutes"
          }
        },
        "required": [
          "limit_id",
          "window",
          "used_percent",
          "remaining_percent",
          "window_minutes",
          "resets_at"
        ],
        "title": "UsageWindowView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Dovet local API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/connections/codex/usage": {
      "get": {
        "operationId": "codex_usage_api_v1_connections_codex_usage_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-dovet-bridge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Bridge"
            }
          },
          {
            "in": "cookie",
            "name": "dovet_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dovet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Codex Usage"
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "run_evidence_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dovet-bridge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Bridge"
            }
          },
          {
            "in": "cookie",
            "name": "dovet_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dovet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEvidenceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evidence"
      }
    },
    "/api/v1/runs/{run_id}/pause": {
      "post": {
        "operationId": "pause_run_api_v1_runs__run_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dovet-csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Csrf"
            }
          },
          {
            "in": "header",
            "name": "x-dovet-bridge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Bridge"
            }
          },
          {
            "in": "cookie",
            "name": "dovet_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dovet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Run"
      }
    },
    "/api/v1/session/issue": {
      "post": {
        "operationId": "issue_pairing_api_v1_session_issue_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dovet-bridge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Bridge"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Issue Pairing Api V1 Session Issue Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Pairing"
      }
    },
    "/api/v1/session/pair": {
      "post": {
        "operationId": "pair_api_v1_session_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Pair Api V1 Session Pair Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair"
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "local_status_api_v1_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-dovet-bridge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dovet-Bridge"
            }
          },
          {
            "in": "cookie",
            "name": "dovet_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dovet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Local Status Api V1 Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Local Status"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}